+++ packages/bootstrap/src/controls/check.tsx
@@ -1,23 +1,25 @@
 import { bound } from "@frui.ts/helpers";
-import { BindingComponent, IBindingProps } from "@frui.ts/views";
-import { Observer } from "mobx-react-lite";
+import { IBindingProps } from "@frui.ts/views";
 import * as React from "react";
 import { Form, FormCheckProps } from "react-bootstrap";
 import { CommonInputProps } from "./commonInputProps";
+import { ValidationControlBase } from "./validationControlBase";
 
-export class Check<TTarget> extends BindingComponent<TTarget, FormCheckProps & CommonInputProps & IBindingProps<TTarget>> {
-  render() {
+export class Check<TTarget> extends ValidationControlBase<TTarget, FormCheckProps & CommonInputProps & IBindingProps<TTarget>> {
+  @bound
+  protected renderInner() {
+    const { noValidation, errorMessage, ...otherProps } = this.inheritedProps;
+    const validationError = this.getValidationError();
+
     return (
-      <Observer>
-        {() => (
-          <Form.Check
-            id={this.props.property}
-            {...this.inheritedProps}
-            checked={!!this.value}
-            onChange={this.handleValueChanged}
-          />
-        )}
-      </Observer>
+      <Form.Check
+        id={this.props.property}
+        {...otherProps}
+        checked={!!this.value}
+        onChange={this.handleValueChanged}
+        isInvalid={!!validationError}
+        feedback={validationError}
+      />
     );
   }